# src/tests/contract/schemas/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Container Management API",
    "version": "1.0.0",
    "description": "API completa para gerenciamento de containers, averbações, usuários e seguradoras",
    "contact": {
      "name": "Suporte API",
      "email": "suporte@container.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "idUsuario": {
            "type": "integer",
            "description": "ID único do usuário"
          },
          "nomeCompleto": {
            "type": "string",
            "description": "Nome completo do usuário"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do usuário"
          },
          "telefone": {
            "type": "string",
            "description": "Telefone do usuário"
          },
          "celular": {
            "type": "string",
            "description": "Celular do usuário"
          },
          "status": {
            "type": "string",
            "enum": [
              "ativo",
              "inativo",
              "bloqueado"
            ],
            "description": "Status do usuário"
          },
          "perfil": {
            "$ref": "#/components/schemas/PerfilUsuario"
          }
        }
      },
      "PerfilUsuario": {
        "type": "object",
        "properties": {
          "idPerfil": {
            "type": "integer"
          },
          "nomePerfil": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "nivelAcesso": {
            "type": "integer"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário"
          },
          "senha": {
            "type": "string",
            "description": "Senha do usuário"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/Usuario"
              },
              "token": {
                "type": "string",
                "description": "JWT token para autenticação"
              }
            }
          }
        }
      },
...
      "Historico": {
        "type": "object",
        "properties": {
          "idHistorico": {
            "type": "integer",
            "description": "ID único do histórico"
          },
          "tabela": {
            "type": "string",
            "description": "Nome da tabela alterada"
          },
          "operacao": {
            "type": "string",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE"
            ],
            "description": "Tipo de operação realizada"
          },
          "registroId": {
            "type": "integer",
            "description": "ID do registro alterado"
          },
          "dadosAnteriores": {
            "type": "object",
            "description": "Dados antes da alteração"
          },
          "dadosNovos": {
            "type": "object",
            "description": "Dados após a alteração"
          },
          "usuarioId": {
            "type": "integer",
            "description": "ID do usuário que realizou a alteração"
          },
          "usuario": {
            "type": "object",
            "properties": {
              "nomeCompleto": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          },
          "dataAlteracao": {
            "type": "string",
            "format": "date-time",
            "description": "Data e hora da alteração"
          },
          "ip": {
            "type": "string",
            "description": "IP do usuário"
          },
          "userAgent": {
            "type": "string",
            "description": "User agent do navegador"
          }
        }
      },
...
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer"
          },
          "itemsPerPage": {
            "type": "integer"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Login de usuário",
        "description": "Realiza o login do usuário e retorna um token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "admin@exemplo.com",
                "senha": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/historico": {
      "get": {
        "tags": [
          "Histórico"
        ],
        "summary": "Obter histórico de alterações",
        "description": "Lista o histórico de alterações do sistema com filtros e paginação",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "tabela",
            "in": "query",
            "description": "Filtrar por tabela",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "operacao",
            "in": "query",
            "description": "Filtrar por operação",
            "schema": {
              "type": "string",
              "enum": [
                "CREATE",
                "UPDATE",
                "DELETE"
              ]
            }
          },
          {
            "name": "usuario",
            "in": "query",
            "description": "Filtrar por usuário",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataInicio",
            "in": "query",
            "description": "Data de início do filtro",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dataFim",
            "in": "query",
            "description": "Data de fim do filtro",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico obtido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "historico": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Historico"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationResponse"
                        },
                        "filtros": {
                          "type": "object",
                          "properties": {
                            "tabela": {
                              "type": "string"
                            },
                            "operacao": {
                              "type": "string"
                            },
                            "usuario": {
                              "type": "string"
                            },
                            "periodo": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}